// packages/www/src/types/playoffs.ts
export interface IPlayoffUserStanding {
  id: number,
  username: string | null,
  picks: {
    value: number;
    pick: {
      team: {
        teamName: string,
        logo: string | null
      } | null,
      seriesWins: number,
      season: string,
      round: number
    }
  }[]
  points: number
}